% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{absorbed_downwelling_longwave_radiation}
\alias{absorbed_downwelling_longwave_radiation}
\title{Downwelling longwave radiation. This essentially takes the same approach
as the version from Wang and Leuning implemented above but only calculates
the downwelling component.}
\usage{
absorbed_downwelling_longwave_radiation(
  Ta,
  l,
  L,
  kb,
  ea,
  ef = 0.95,
  es = 0.9,
  kd = 0.8,
  aslist = FALSE
)
}
\arguments{
\item{Ta}{Air temperature (K)}

\item{l}{Cumulative leaf area index from top of canopy (m2 / m2)}

\item{L}{Total leaf area index for site}

\item{kb}{Attenuation coefficient for black leaves under beam radiation}

\item{ea}{Effective emissivity of the sky}

\item{ef}{Emissivity of foliage}

\item{es}{Emissivity of the soil}

\item{kd}{Attenuation coefficient for black leaves under diffuse radiation}
}
\value{
L_dn, downwelling longwave radiation
}
\description{
Downwelling longwave radiation. This essentially takes the same approach
as the version from Wang and Leuning implemented above but only calculates
the downwelling component.
}
\keyword{internal}
