% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{absorbed_longwave_radiation}
\alias{absorbed_longwave_radiation}
\title{Absorbed shortwave radiation within a tree canopy. Comes from Wang and
Leuning (1998) Eqs. B8-11}
\usage{
absorbed_longwave_radiation(
  Ta,
  l,
  L,
  kb,
  ea,
  ef = 0.95,
  es = 0.9,
  kd = 0.8,
  aslist = FALSE
)
}
\arguments{
\item{Ta}{Air temperature (K)}

\item{l}{Cumulative leaf area index from top of canopy (m2 / m2)}

\item{L}{Total leaf area index for site}

\item{kb}{Attenuation coefficient for black leaves under beam radiation}

\item{ea}{Effective emissivity of the sky}

\item{ef}{Emissivity of foliage}

\item{es}{Emissivity of the soil}

\item{kd}{Attenuation coefficient for black leaves under diffuse radiation}

\item{aslist}{Return partitioned sun/shade fluxes as a list?}
}
\value{
q_total, total absorbed longwave radiation of sunlit and
shaded leaves (W/m2)
}
\description{
Absorbed shortwave radiation within a tree canopy. Comes from Wang and
Leuning (1998) Eqs. B8-11
}
\keyword{internal}
