% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{absorbed_shortwave_radiation}
\alias{absorbed_shortwave_radiation}
\title{Absorbed shortwave radiation within a tree canopy. Comes from Wang and
Leuning (1998) Eqs. B2-3}
\usage{
absorbed_shortwave_radiation(
  SW_beam,
  SW_dif,
  l,
  kb_real,
  kd_real,
  kb,
  kd = 0.8,
  s = 0.5,
  rho_b = 0.1,
  rho_d = 0.1,
  aslist = FALSE
)
}
\arguments{
\item{SW_beam}{Top-of-canopy direct shortwave radiation (W/m2)}

\item{SW_dif}{Top-of-canopy diffuse shortwave radiation (W/m2)}

\item{l}{Cumulative leaf area index from top of canopy (m2 / m2)}

\item{kb_real}{Empirical attenuation coefficient under direct radiation}

\item{kd_real}{Empirical attenuation coefficient under diffuse radiation}

\item{kb}{Attenuation coefficient for black leaves under direct radiation}

\item{kd}{Attenuation coefficient for black leaves under diffuse radiation}

\item{s}{Leaf scattering coefficient = 1 - absorptance (0-1)}

\item{rho_b}{Canopy reflectance coefficient for direct radiation}

\item{rho_d}{Canopy reflectance coefficient for diffuse radiation}

\item{aslist}{Return partitioned sun/shade fluxes as a list?}
}
\value{
q_total, total absorbed shortwave radiation of sunlit and
shaded leaves (W/m2)
}
\description{
Absorbed shortwave radiation within a tree canopy. Comes from Wang and
Leuning (1998) Eqs. B2-3
}
\keyword{internal}
