% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tower_util.R
\name{amf_var_heights}
\alias{amf_var_heights}
\title{Return Ameriflux variable info (possibly from cache). This function only exists so that it can be memoized.}
\usage{
amf_var_heights()
}
\value{
see \link[amerifluxr]{amf_var_info}
}
\description{
Return Ameriflux variable info (possibly from cache). This function only exists so that it can be memoized.
}
\keyword{internal}
