% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{conductance_boundary_heat_broadleaf}
\alias{conductance_boundary_heat_broadleaf}
\title{Leaf boundary layer conductance to sensible heat transfer in
broad leaves.}
\usage{
conductance_boundary_heat_broadleaf(u, d, rho_mol)
}
\arguments{
\item{u}{Wind speed (m/sec)}

\item{d}{Characteristic leaf dimension (m)}

\item{rho_mol}{Molar air density (mol / m^3)}
}
\value{
gbH, conductance to sensible heat transfer (m/sec)
}
\description{
Leaf boundary layer conductance to sensible heat transfer in
broad leaves.
}
\keyword{internal}
