% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{decoupling_coefficient}
\alias{decoupling_coefficient}
\title{Air-leaf temperature decoupling coefficient (Omega). Range is 0-1. When
omega is near 0, boundary layer conductance is high and leaf temperature
is well-coupled to air temperature. When omega is 1, boundary layer
conductance is low and latent heat flux is radiation-limited.}
\usage{
decoupling_coefficient(gR, gbH, gs, d_esat, gamma, a_lw = 0.95)
}
\arguments{
\item{gR}{Conductance to radiative heat transfer (mol / m2 sec)}

\item{gbH}{Boundary layer conductance to sensible heat (mol / m2 sec)}

\item{gs}{Stomatal conductance (mol / m2 sec)}

\item{d_esat}{Saturation VP vs. air temperature slope (kPa / K)}

\item{gamma}{Psychrometric constant (kPa / K)}

\item{a_lw}{Longwave absorptance}
}
\value{
omega, decoupling coefficient.
}
\description{
Air-leaf temperature decoupling coefficient (Omega). Range is 0-1. When
omega is near 0, boundary layer conductance is high and leaf temperature
is well-coupled to air temperature. When omega is 1, boundary layer
conductance is low and latent heat flux is radiation-limited.
}
\keyword{internal}
