% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{dry_air_density}
\alias{dry_air_density}
\title{Density of dry air.}
\usage{
dry_air_density(Ta, Pa, cp = eb_constants_$cp, M = eb_constants_$Mair)
}
\arguments{
\item{Ta}{Air temperature (K)}

\item{Pa}{Air pressure (kPa)}

\item{cp}{Heat capacity of dry air (J/kg K)}

\item{M}{Molar mass of dry air (kg / mol)}
}
\value{
rho, air density (kg / m3)
}
\description{
Density of dry air.
}
\keyword{internal}
