% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{dry_air_molar_density}
\alias{dry_air_molar_density}
\title{Molar density of dry air.}
\usage{
dry_air_molar_density(Ta, Pa, cp = eb_constants_$cp, M = eb_constants_$Mair)
}
\arguments{
\item{Ta}{Air temperature (K)}

\item{Pa}{Air pressure (Pa)}

\item{cp}{Heat capacity of dry air (J/kg K)}

\item{M}{Molar mass of dry air (kg/mol)}
}
\value{
rho_mol, molar density of dry air (mol / m3)
}
\description{
Molar density of dry air.
}
\keyword{internal}
