% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_eb_numeric.R
\name{energy_balance_error}
\alias{energy_balance_error}
\alias{energy_balance_driver}
\title{Numerical leaf energy balance model}
\usage{
energy_balance_error(
  Tl,
  Ta,
  u,
  gs,
  Pa,
  RH,
  SW_abs,
  LW_abs_dn,
  a_lw = 0.95,
  a_sw = 0.5,
  d = 0.01,
  aslist = FALSE
)

energy_balance_driver(
  Ta,
  u,
  gs,
  Pa,
  RH,
  SW_abs,
  LW_abs_dn,
  a_lw = 0.95,
  a_sw = 0.5,
  d = 0.01,
  bounds = 20,
  run_no_transp = TRUE
)
}
\arguments{
\item{Tl}{Leaf temperature (K)}

\item{Ta}{Air temperature (K)}

\item{u}{Wind speed (m/s)}

\item{gs}{Stomatal conductance (mol / m2 sec)}

\item{Pa}{Air pressure (kPa)}

\item{RH}{Relative humidity (\%)}

\item{SW_abs}{Absorbed shortwave radiation (W/m2)}

\item{LW_abs_dn}{Absorbed downwelling longwave radiation (W/m2)}

\item{a_lw}{Longwave absorptance (0-1)}

\item{a_sw}{Shortwave absorptance (0-1)}

\item{d}{Leaf characteristic dimension (m)}

\item{aslist}{Return error and fluxes as a list? If false, only returns error.}

\item{bounds}{Deviation from air temperature to define the search interval. For example, if Ta = 290 K and bounds = 10 K, the model will search for an optimal leaf temperature in the range of \link{280, 300} K.}

\item{run_no_transp}{Should the model be run a second time with gs = 0.001? This is necessary to calculate temperature forcings from the fluxes. If TRUE, additional outputs are returned.}
}
\value{
If aslist == TRUE, a list containing calculated heat fluxes, conductances, and decoupling. Otherwise, just the energy balance error.
}
\description{
Numerical leaf energy balance model
}
\section{Functions}{
\itemize{
\item \code{energy_balance_driver()}: User-friendly driver function to run
numerical optimization.

}}
\keyword{internal}
