% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{esat_slope}
\alias{esat_slope}
\title{Slope of saturation vapor pressure with respect to temperature. Found by
differentiating Tetens's formula.}
\usage{
esat_slope(Ta)
}
\arguments{
\item{Ta}{Air tempertaure (K)}
}
\value{
desat, Slope of esat w.r.t. air temperature (kPa / K)
}
\description{
Slope of saturation vapor pressure with respect to temperature. Found by
differentiating Tetens's formula.
}
\keyword{internal}
