% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_aq_curves.R
\name{fit_aq_curves}
\alias{fit_aq_curves}
\title{Fit AQ curves to whole-canopy photosynthesis}
\usage{
fit_aq_curves(site_meta, site_flux_qc, tower_dir, outdir)
}
\arguments{
\item{site_meta}{Site metadata table \code{\link{site_meta}}.}

\item{site_flux_qc}{Partitioned, QC'd flux data generated by \code{\link[=partition_neon_flux]{partition_neon_flux()}}.}

\item{tower_dir}{Directory where tower data files were written by \code{\link[=get_amf_tower_data]{get_amf_tower_data()}}.}

\item{outdir}{Directory where AQ curve parameter table will be written.}
}
\description{
To disaggregate whole-canopy photosynthesis, we assume that each canopy layer's contribution to the whole-canopy value is proportional to the radaition flux reachin gthe slice. But, photosynthesis saturates under a certain amount of radiation. We estimate this threshold by comparing top-of-canopy shortwave flux with whole-canopy GPP.
}
