% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_medlyn_slopes.R
\name{fit_medlyn_slopes}
\alias{fit_medlyn_slopes}
\title{Fit Medlyn stomatal conductance model to each site}
\usage{
fit_medlyn_slopes(
  site_meta,
  site_lai,
  site_flux_qc,
  tower_dir,
  outdir,
  rain_filter = 3
)
}
\arguments{
\item{site_meta}{Site metadata table \code{\link{site_meta}}.}

\item{site_lai}{Site leaf area index table \code{\link{modis_lai}}.}

\item{site_flux_qc}{Partitioned, QC'd eddy covariance flux table generated by \code{\link[=partition_neon_flux]{partition_neon_flux()}}.}

\item{tower_dir}{Directory where tower data was written by \code{\link[=get_amf_tower_data]{get_amf_tower_data()}}.}

\item{outdir}{Output directory where table of constants will be written.}

\item{rain_filter}{Number of days after a rain event where observations are discarded. This is necessary to ensure that transpiration, not evaporation, is driving the latent heat flux signal.}
}
\description{
We use a Medlyn model to link canopy photosynthesis with stomatal conductance and latent heat flux. This model requires a slope parameter to work. This function calculates the slope parameter for each site by inverting the Penman-Monteith equation as in Knauer et al. (2018), GCB.
}
