% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_amf_tower_data.R
\name{get_amf_tower_data}
\alias{get_amf_tower_data}
\title{Download tower data from Ameriflux}
\usage{
get_amf_tower_data(site_meta, outdir, user, email)
}
\arguments{
\item{site_meta}{Site metadata table \code{\link{site_meta}}.}

\item{outdir}{Directory to write tower CSVs. A subfolder named \code{amf_tower_data} will be created here.}

\item{user}{Ameriflux username.}

\item{email}{Ameriflux email.}
}
\description{
This function downloads the current Ameriflux BASE product for each site listed in \code{site_meta}. It also isolates all \code{T_CANOPY} observations and writes these to a separate table. For the 11 sites listed in \code{\link{site_meta}}, about 2GB of storage is needed.
}
