% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neon_flux.R
\name{get_neon_flux}
\alias{get_neon_flux}
\title{Download eddy covariance data from NEON}
\usage{
get_neon_flux(site_meta, token, outdir)
}
\arguments{
\item{site_meta}{Site metadata table \code{\link{site_meta}}}

\item{token}{NEON API token. Can be NULL or NA if you don't have one.}

\item{outdir}{Directory to write stacked flux tables.}
}
\description{
This function downloads eddy covariance data from NEON, stacks the level 4 NSAE data and QC flags, and then writes that to a CSV in \code{outdir}. For growing season data only, this downloads about 60 GB total, but uses <500 MB on disk at any one time by deleting intermediate files. This is suboptimal, but is currently recommended for large flux downloads (see \href{https://github.com/NEONScience/NEON-utilities/issues/131}{this github issue}).

For reproducibility with the manuscript, this script also excludes any flux products after 2021.
}
