% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neon_lidar.R
\name{get_neon_lidar}
\alias{get_neon_lidar}
\title{Download NEON aerial LiDAR near flux towers}
\usage{
get_neon_lidar(site_meta, outdir)
}
\arguments{
\item{site_meta}{Site metadata table \code{\link{site_meta}}.}

\item{outdir}{Output directory where processed point clouds are saved.}
}
\description{
This function downloads all aerial LiDAR tiles within 50 m of the coordinates in \code{site_meta}, normalizes height, and then clips the points together into one 50 m circle.
}
