% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tower_util.R
\name{get_var_heights}
\alias{get_var_heights}
\title{Get variables matching a regex and their height of observation.}
\usage{
get_var_heights(all_vars, regex, site)
}
\arguments{
\item{all_vars}{Character vector of variables in a tower data frame.}

\item{regex}{Character string used to \code{grep} variables out of \code{all_vars}.}

\item{site}{Site code used to filter results from \code{amf_var_info}.}
}
\value{
a list containing the regex'd varnames and their heights
}
\description{
Get variables matching a regex and their height of observation.
}
\keyword{internal}
