% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_within_canopy_meteorology.R
\name{get_within_canopy_meteorology}
\alias{get_within_canopy_meteorology}
\title{Interpolate within-canopy meteorology to model layers}
\usage{
get_within_canopy_meteorology(site_meta, site_lai, site_lad, site_flux, outdir)
}
\arguments{
\item{site_meta}{Site metadata table \code{\link{site_meta}}}

\item{site_lai}{Site leaf area index table \code{\link{manual_lai}}}

\item{site_lad}{Site leaf area density profiles generated by \code{\link[=fit_neon_lidar_k]{fit_neon_lidar_k()}}.}

\item{site_flux}{Partitioned eddy covariance flux generated by \code{\link[=partition_neon_flux]{partition_neon_flux()}}.}

\item{outdir}{Output directory where interpolated meteorology table will be written.}
}
\description{
In all sites, model layers are not necessarily at the same physical height as meteorological sensors. This is because model layers are based on cumulative LAI, which does not change linearly with height. Therefore, we use linear interpolation to approximate microclimate conditions at the position of each model layer. This script looks up the sensor heights of relevant variabels in the tower data and then interpolates to the layer positions.
}
