% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_gbh_sensitivity.R
\name{gs_gbh_sensitivity}
\alias{gs_gbh_sensitivity}
\title{Stomatal/boundary layer sensitivity analysis}
\usage{
gs_gbh_sensitivity(
  outdir,
  gs_min = 0.01,
  gs_max = 0.5,
  gbH_min = 0.75,
  gbH_max = 5,
  Ta = 293,
  Pa = 100,
  RH = 50,
  a_lw = 0.98,
  a_sw = 0.5
)
}
\arguments{
\item{outdir}{Directory where model results will be written.}

\item{gs_min}{Minimum value for stomatal conductance (mol m-2 s-1).}

\item{gs_max}{Maximum value for stomatal conductance (mol m-2 s-1).}

\item{gbH_min}{Minimum value for boundary layer conductance (mol m-2 s-1).}

\item{gbH_max}{Maximum value for boundary layer conductance (mol m-2 s-1).}

\item{Ta}{Air temperature (K).}

\item{Pa}{Air pressure (kPa).}

\item{RH}{Relative humidity (\%, from 0-100).}

\item{a_lw}{Longwave absorptivity of leaves.}

\item{a_sw}{Shortwave absorptivity of leaves.}
}
\description{
Stomatal conductance and boundary layer conductance are two important parameters that drive the temperature of a leaf. This function evaluates how these parameter drive temperature under a low, medium, and high radiation load.
}
