% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tower_util.R
\name{interpolate_tower_data}
\alias{interpolate_tower_data}
\title{Interpolate flux tower micrometeorology data at arbitrary z positions.}
\usage{
interpolate_tower_data(tower, vars, var_z, interp_z, name_prefix = "")
}
\arguments{
\item{tower}{Data frame containing relevant tower observations.}

\item{vars}{Vector of names in \code{tower} to interpolate.}

\item{var_z}{Vector of heights above ground where \code{vars} are measured.}

\item{interp_z}{Vector of heights above ground at which to interpolate \code{vars}.}

\item{name_prefix}{String prepended to names of the resulting data frame,
which by default are \code{interp_z}.}
}
\value{
Dataframe with the observations in \code{tower} interpolated at \code{interp_z}
}
\description{
Interpolate flux tower micrometeorology data at arbitrary z positions.
}
