% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{latent_heat}
\alias{latent_heat}
\title{Actual latent heat flux. Calculated as mixture of stomtally-imposed and
equilibrium endpoints. See Still et al. (2022) for details.}
\usage{
latent_heat(omega, le_eq, le_imp)
}
\arguments{
\item{omega}{Decoupling coefficient (0-1)}

\item{le_eq}{Equilibrium endpoint (W/m2)}

\item{le_imp}{Stomatally-imposed endpoint (W/m2)}
}
\value{
le, total latent heat flux (W/m2)
}
\description{
Actual latent heat flux. Calculated as mixture of stomtally-imposed and
equilibrium endpoints. See Still et al. (2022) for details.
}
\keyword{internal}
