% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{latent_heat_equilibrium}
\alias{latent_heat_equilibrium}
\title{Equilibrium latent heat flux. This is the flux when leaf temperature is
decoupled from air temperature.}
\usage{
latent_heat_equilibrium(Rn, d_esat, gamma, gR, gbH)
}
\arguments{
\item{Rn}{Net isothermal radiation (W / m2)}

\item{d_esat}{Slope of saturation VP vs. temperature (kPa / K)}

\item{gamma}{Psychrometric constant (kPa / K)}

\item{gR}{Leaf conductance to radiative heat transfer (mol / m2 sec)}

\item{gbH}{Leaf boundary layer conductance to heat transfer (mol / m2 sec)}
}
\value{
LE_eq, equilibrium latent heat flux (W / m2)
}
\description{
Equilibrium latent heat flux. This is the flux when leaf temperature is
decoupled from air temperature.
}
\keyword{internal}
