% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeafTemperature-package.R
\docType{package}
\name{LeafTemperature-package}
\alias{LeafTemperature}
\alias{LeafTemperature-package}
\title{LeafTemperature: Modeling Leaf Temperature in Evergreen Forests}
\description{
This package archives analysis scripts and code to generate figures for our leaf temperature modeling manuscript.
}
\author{
\strong{Maintainer}: Keenan Ganz \email{ganzk@uw dot edu} (\href{https://orcid.org/0000-0002-8486-3959}{ORCID})

}
\keyword{internal}
