% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{li_gs}
\alias{li_gs}
\title{Generalization of Medlyn model of stomatal conductance from Li et al. (2019).}
\usage{
li_gs(A, VPD_l, CO2, g1, g0 = 0, m = 0.5)
}
\arguments{
\item{A}{Net photosynthesis (umol CO2 / m^2 s)}

\item{VPD_l}{Ecosystem VPD (kPa, see reference)}

\item{CO2}{Atmospheric CO2 concentration (umol CO2 / mol air == ppm)}

\item{g1}{Slope parameter (kPa^0.5, usually in range 1-10)}

\item{g0}{Intercept parameter (mol / m^2 s, usually set to 0)}

\item{m}{VPD exponent parameter}
}
\value{
gs, stomatal conductance (mol m-2 s-1)
}
\description{
Generalization of Medlyn model of stomatal conductance from Li et al. (2019).
}
