% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeafTemperature-package.R
\docType{data}
\name{manual_lai_images}
\alias{manual_lai_images}
\alias{manual_lai}
\title{Site leaf area index tables}
\description{
We used a random sample of NEON DHPs to calculate leaf area index in each of our study sites. \code{manual_lai_images} documents what that random sample was for reproducibility, and \code{manual_lai} records the median LAI we calculated at each site using the \code{hemispheR} package. See \code{get_neon_lai.R} for more information.
}
