% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{medlyn_gs}
\alias{medlyn_gs}
\title{Empirical model of stomatal conductance based on Medlyn et al. (2011).}
\usage{
medlyn_gs(A, VPD, CO2, g1, g0 = 0)
}
\arguments{
\item{A}{Net photosynthesis (umol CO2 / m^2 s)}

\item{VPD}{Vapor pressure deficit (kPa)}

\item{CO2}{Atmospheric CO2 concentration (umol CO2 / mol air == ppm)}

\item{g1}{Slope parameter (kPa^0.5, usually in range 1-10)}

\item{g0}{Intercept parameter (mol / m^2 s, usually set to 0)}
}
\value{
gs, stomatal conductance (mol m-2 s-1)
}
\description{
Empirical model of stomatal conductance based on Medlyn et al. (2011).
}
\keyword{internal}
