% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_radiometer_comparison.R
\name{model_radiometer_comparison}
\alias{model_radiometer_comparison}
\title{Validate energy balance model by comparing with tower radiometers}
\usage{
model_radiometer_comparison(eb_result, rad_tcan, outdir)
}
\arguments{
\item{eb_result}{Energy balance model output generated by \code{\link[=run_neon_energy_balance]{run_neon_energy_balance()}}.}

\item{rad_tcan}{Table of radiometer canopy temperature data generated by \code{\link[=get_amf_tower_data]{get_amf_tower_data()}}.}

\item{outdir}{Directory where model/radiometer comparison data will be written.}
}
\description{
Tower radiometers measure canopy biological temperature, which is similar to leaf temperature but slightly different. It's the best validation data available at all our study sites. But, like with the meteorology data, radiometers aren't at the same positions as the model layers. This function generates a crosswalk between model layers and radiometers so we compare each model with the nearest radiometer. If multiple model layers correspond to a single radiometer, temperature across all layers are averaged and compared with the radiometer.
}
