% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeafTemperature-package.R
\docType{data}
\name{modis_lai}
\alias{modis_lai}
\title{MODIS leaf area index table}
\description{
Due to large variation in DHP-derived LAI, we decided to use LAI from MODIS. The process for generating these data was as follows. First, we used AppEARS to derive a point sample of all 8-day 500-m MODIS LAI images from May - September, 2018 - 2022. This request is documented under appears_request in the package directory (or under inst, if you are looking on GitHub). Then, we subset the AppEARS request to those observations where the QC band was zero and calculated site LAI as the 95th percentile of the LAI band. This procedure resulted in better agreement with literature values for LAI than the DHP method, which guided our decision.
}
