% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{net_isothermal_radiation}
\alias{net_isothermal_radiation}
\title{Net isothermal radiation.}
\usage{
net_isothermal_radiation(SW_abs, LW_abs_dn, Ta, a_lw)
}
\arguments{
\item{Ta}{Air temperature (K)}

\item{a_lw}{Absorptivity of longwave radiation (0.95)}

\item{SW_dir}{Direct downwelling shortwave radiation (W/m2)}

\item{SW_dif}{Diffuse downwelling shortwave radiation (W/m2)}

\item{SW_out}{Upwelling shortwave radiation (W/m2)}

\item{LW_down}{Downwelling longwave radiation (W/m2)}

\item{is_sunlit}{Is the leaf in shade? (1/0)}

\item{illuminance}{Accessibility of diffuse light (0-1)}

\item{a_sw}{Absorptivity of shortwave radiation (0.5)}
}
\value{
Rn, Net isothermal radiation (W/m2)
}
\description{
Net isothermal radiation.
}
\keyword{internal}
