% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{outgoing_longwave_radiation}
\alias{outgoing_longwave_radiation}
\title{Outgoing (upwelling) longwave radiation from a leaf. This version just models
the leaf as a blackbody since the upwelling component from Wang and Leuning
(1998) is essentially equivalent to this. The flux is doubled because both
sides of the leaf participate in heat transfer.}
\usage{
outgoing_longwave_radiation(Tl, ef = 0.95)
}
\arguments{
\item{Tl}{Leaf temperature (K)}

\item{ef}{Effective foliage emissivity (0-1)}
}
\value{
L_up, upwelling longwave radiation
}
\description{
Outgoing (upwelling) longwave radiation from a leaf. This version just models
the leaf as a blackbody since the upwelling component from Wang and Leuning
(1998) is essentially equivalent to this. The flux is doubled because both
sides of the leaf participate in heat transfer.
}
\keyword{internal}
