% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_neon_flux.R
\name{partition_neon_flux}
\alias{partition_neon_flux}
\title{Run REddyProc on NEON eddy covariance data}
\usage{
partition_neon_flux(site_meta, raw_flux, tower_dir, out_dir)
}
\arguments{
\item{site_meta}{Site metadata table \code{\link{site_meta}}.}

\item{raw_flux}{Unprocessed flux table generated by \code{\link[=get_neon_flux]{get_neon_flux()}}.}

\item{tower_dir}{Directory containing tower data files generated by \code{\link[=get_amf_tower_data]{get_amf_tower_data()}}.}

\item{out_dir}{Output directory where partitioned flux will be written.}
}
\description{
This function groups raw NEON flux data by site-year and passes each one to REddyProc for partitioning into gross primary production and ecosystem respiration. Due to data limitations, we provide all data to REddyProc. Once partitioned, we drop observations with any of the NEON QC flags set.
}
