% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{psychrometric_constant}
\alias{psychrometric_constant}
\title{Psychrometric constant - relates partial pressure of water to
air temperature.}
\usage{
psychrometric_constant(
  Pa,
  cp = eb_constants_$cp,
  Mwr = eb_constants_$MWr,
  l = eb_constants_$l
)
}
\arguments{
\item{Pa}{Air pressure (kPa)}

\item{cp}{Heat capacity of dry air (J/kg K)}

\item{Mwr}{Ratio of molar mass of water to molar mass of dry air}

\item{l}{Heat of vaporization of water (J / mol)}
}
\value{
gamma, Psycrometric constant (kPa / K)
}
\description{
Psychrometric constant - relates partial pressure of water to
air temperature.
}
\keyword{internal}
