% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Re-run leaf temperature modeling pipeline}
\usage{
run_analysis(
  amf_user,
  amf_email,
  neon_token,
  work_dir,
  log = NULL,
  skip = "nnnnnnnnnnn"
)
}
\arguments{
\item{amf_user}{Ameriflux username for data download}

\item{amf_email}{Ameriflux email for data download}

\item{neon_token}{Token for NEON API. Can be NA, but this will makes your downloads take longer.}

\item{work_dir}{Working directory to write tables.}

\item{log}{If not \code{NULL}, all output will be written to this file.}

\item{skip}{11-character string describing whether you want to skip certain analysis steps. This can be useful if you hit an error halfway through and want to skip all the steps that succeeded initially.}
}
\description{
Re-run leaf temperature modeling pipeline
}
\details{
This function re-runs the leaf temperature modeling pipeline we describe in the manuscript. It begins by downloading ~2 GB of data to your machine from Ameriflux and NEON. If you run the whole thing, you should expect it to take 2-3 hours. Most of that time is just downloading data, the actual modeling part takes just a few minutes. We have done our best to make the code robust, but ti is always possible that the analysis breaks halfway through. If that happens, you can adjust the \code{skip} parameter to ignore the steps that have already completed.

The skip parameter is an 11-character string. If the i-th character is 'y', then the i-th analysis step is skipped. If a step is skipped, it is assumed that the relevant data has already been written to \code{work_dir} with the expected filename. The order of steps is as follows:
\enumerate{
\item \code{\link[=get_amf_tower_data]{get_amf_tower_data()}}
\item \code{\link[=get_neon_flux]{get_neon_flux()}}
\item \code{\link[=get_neon_lidar]{get_neon_lidar()}}
\item \code{\link[=fit_neon_lidar_k]{fit_neon_lidar_k()}}
\item \code{\link[=partition_neon_flux]{partition_neon_flux()}}
\item \code{\link[=get_within_canopy_meteorology]{get_within_canopy_meteorology()}}
\item \code{\link[=fit_aq_curves]{fit_aq_curves()}}
\item \code{\link[=fit_medlyn_slopes]{fit_medlyn_slopes()}}
\item \code{\link[=run_neon_energy_balance]{run_neon_energy_balance()}}
\item \code{\link[=model_radiometer_comparison]{model_radiometer_comparison()}}
\item \code{\link[=gs_gbh_sensitivity]{gs_gbh_sensitivity()}}
}

See the linked functions for more information. Once the analysis finishes, you can run \code{\link[=write_all_figures]{write_all_figures()}} to generate the figures we show in the manuscript.
}
