% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_neon_energy_balance.R
\name{run_neon_energy_balance}
\alias{run_neon_energy_balance}
\title{Run the final energy balance model}
\usage{
run_neon_energy_balance(
  site_meta,
  site_lai,
  site_flux_qc,
  aq_constants,
  medlyn_constants,
  lidar_constants,
  interp_meteo,
  tower_dir,
  outdir
)
}
\arguments{
\item{site_meta}{Site metadata table \code{\link{site_meta}}.}

\item{site_lai}{Site leaf area index table \code{\link{manual_lai}}.}

\item{site_flux_qc}{Partitioned, QC'd eddy covariance flux table generated by \code{\link[=partition_neon_flux]{partition_neon_flux()}}.}

\item{aq_constants}{AQ curve constants generated by \code{\link[=fit_aq_curves]{fit_aq_curves()}}.}

\item{medlyn_constants}{Medlyn slope parameters generated by \code{\link[=fit_medlyn_slopes]{fit_medlyn_slopes()}}.}

\item{lidar_constants}{Light attenuation constants generated by \code{\link[=fit_neon_lidar_k]{fit_neon_lidar_k()}}.}

\item{interp_meteo}{Interpolated meteorlogical measurements generated by \code{\link[=interpolate_canopy_meteorology]{interpolate_canopy_meteorology()}}.}

\item{tower_dir}{Directory where tower data was written by \code{\link[=get_amf_tower_data]{get_amf_tower_data()}}.}

\item{outdir}{Directory where output model artifacts will be written.}
}
\description{
This file runs a numeric energy balance model using all the parameters and data produced by preceding functions. See the manuscript for full details on the model spec. Three model runs are done: one with no modifications to the input data, one with the leaves modeled as 10 cm large instead of 1 cm, and one in WREF only that varies the Medlyn slope parameter from 1 to 9.
}
