% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{saturation_vapor_pressure}
\alias{saturation_vapor_pressure}
\title{Saturation vapor pressure from Tetens's formula.}
\usage{
saturation_vapor_pressure(Ta)
}
\arguments{
\item{Ta}{Air temperature (K)}
}
\value{
esat, saturation vapor pressure (kPa)
}
\description{
Saturation vapor pressure from Tetens's formula.
}
\keyword{internal}
