% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{sensible_heat}
\alias{sensible_heat}
\title{Sensible heat flux between a leaf and its boundary layer. Flux is multiplied
by 2 since both sides of leaf participate in heat transfer.}
\usage{
sensible_heat(gbH, Tl, Ta)
}
\arguments{
\item{gbH}{Leaf boundary layer conductance to heat transfer (mol / m2 sec)}

\item{Tl}{Leaf temperature (K)}

\item{Ta}{Air temperature (K)}
}
\value{
H, sensible heat flux (W/m2)
}
\description{
Sensible heat flux between a leaf and its boundary layer. Flux is multiplied
by 2 since both sides of leaf participate in heat transfer.
}
\keyword{internal}
