% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{total_conductance}
\alias{total_conductance}
\title{Total conductance. Assumes that boundary layer conductance to water vapor
is 1.08 times boundary layer heat conductance.}
\usage{
total_conductance(gs, gbH)
}
\arguments{
\item{gs}{Stomatal conductance (mol / m2 sec)}

\item{gbH}{Boundary layer heat conductance (mol / m2 sec)}
}
\value{
gtot, total conductance (mol / m2 sec)
}
\description{
Total conductance. Assumes that boundary layer conductance to water vapor
is 1.08 times boundary layer heat conductance.
}
\keyword{internal}
