% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{vapor_pressure_deficit}
\alias{vapor_pressure_deficit}
\title{Vapor pressure deficit, the difference between the saturation vapor pressure
and the actual vapor pressure.}
\usage{
vapor_pressure_deficit(Ta, RH)
}
\arguments{
\item{Ta}{Air temperature (K)}

\item{RH}{Relative humidity (\%, 0-100)}
}
\value{
vpd, vapor pressure deficit (kPa)
}
\description{
Vapor pressure deficit, the difference between the saturation vapor pressure
and the actual vapor pressure.
}
\keyword{internal}
