% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{write_all_figures}
\alias{write_all_figures}
\title{Write all manuscript figures to a directory}
\usage{
write_all_figures(site_meta, search_dir, out_dir, overwrite = FALSE)
}
\arguments{
\item{site_meta}{Table of site details, you should use \code{\link{site_meta}} unless you want to use a different set of sites.}

\item{search_dir}{Directory the function will look in for data tables. This should be the same as \code{work_dir} in \code{\link[=run_analysis]{run_analysis()}}.}

\item{out_dir}{Output directory for all the figures.}

\item{overwrite}{If a figure already exists, should it be overwritten?}
}
\value{
Void
}
\description{
Write all manuscript figures to a directory
}
\details{
This function calls all of the \verb{fig*} functions in sequence. It takes care of reading the right tables to reproduce what you see in the manuscript.
}
