% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{conductance_radiative_heat}
\alias{conductance_radiative_heat}
\title{Leaf conductance to radiative heat transfer.}
\usage{
conductance_radiative_heat(
  Ta,
  Pa,
  rho,
  rho_mol,
  a_lw = 0.98,
  cp = eb_constants_$cp,
  sb = eb_constants_$sb
)
}
\arguments{
\item{Ta}{Air temperature (K)}

\item{Pa}{Air pressure (kPa)}

\item{a_lw}{Longwave absorptance}

\item{cp}{Heat capacity of dry air}
}
\value{
gR, leaf conductance to radiative heat transfer (m/sec)
}
\description{
Leaf conductance to radiative heat transfer.
}
\keyword{internal}
